% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_XML.R
\name{convert_XML}
\alias{convert_XML}
\title{Create BEAST XML file}
\usage{
convert_XML(file, template, name, loc = FALSE, chainLength = FALSE,
  storeEvery = FALSE)
}
\arguments{
\item{file}{Input file with ID and sequences. Accepts .nex, .fasta and .phy files}

\item{template}{XML file to be used as a template.}

\item{name}{File name that the resulting XML file will be saved as.}

\item{loc}{Tab delimited file containing location data.}

\item{chainLength}{Integer determining the chainlength of the MCMC run. Defaults to template file specifications.}

\item{storeEvery}{Integer determining how often to store the results of the MCMC run. Defaults to template file specifications.}
}
\description{
Import sequences and place them in an XML file with the required settings
}
