% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_XML.R
\name{template_XML}
\alias{template_XML}
\title{Create template XML file}
\usage{
template_XML(template, file)
}
\arguments{
\item{template}{Name of template XML file}

\item{file}{Name of file to write output template XML file to}
}
\description{
Take a template XML file and remove alignment data
}
