% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_XML.R
\name{convert_XML}
\alias{convert_XML}
\title{Create BEAST XML file}
\usage{
convert_XML(file, datatype, template, name, dates = FALSE, loc = FALSE,
  chainLength = FALSE, storeEvery = FALSE)
}
\arguments{
\item{file}{Input file with ID and sequences. Accepts .nex, .fasta and .phy files}

\item{datatype}{Type of date in the sequence input file. Either "nucleotide", "aminoacid", or "binary".}

\item{template}{XML file to be used as a template.}

\item{name}{File name that the resulting XML file will be saved as.}

\item{dates}{Tab delimited file containing tip date data. Date defaults to zero if none supplied}

\item{loc}{Tab delimited file containing location data.}

\item{chainLength}{Integer determining the length of the MCMC chain. Defaults to template file specifications.}

\item{storeEvery}{Integer determining how often to store the state. Defaults to template file specifications.}
}
\description{
Import sequences and places them in an XML file with the required settings
}
