% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_XML.R
\name{convert_XML}
\alias{convert_XML}
\title{Create BEAST XML file}
\usage{
convert_XML(file, datatype, template, name, dates = NULL, location = NULL,
  chainLength = NULL, storeEvery = NULL)
}
\arguments{
\item{file}{The name of the input file containing the alignment. Works with
.nex, .fasta and .phy files}

\item{datatype}{The type of data used in the alignment input file. Either
'nucleotide', 'aminoacid', or 'binary'.}

\item{template}{The name of the BEAST input XML file to be used as a
template. This file can be produced by BEAUti}

\item{name}{The name of the file that the resulting XML file will be written
to.}

\item{dates}{The name of the tab delimited file containing tip date data.}

\item{location}{The name of the tab delimited file containing location data.}

\item{chainLength}{An integer determining the length of the MCMC chain. If
not stated will stay as the value specified in the original template XML
file.}

\item{storeEvery}{An integer determining how often to store the state. If not
stated will stay as the value specified in the original template XML file.}
}
\description{
Imports sequences from a file and places them in an XML file
  generated by BEAUti for use with BEAST.
}
