% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_sequences.R
\name{import_sequences}
\alias{import_sequences}
\title{Import Sequences}
\usage{
import_sequences(file)
}
\arguments{
\item{file}{Input file containing sequences. Accepts .nex, .fasta and .phy files}
}
\description{
Imports a file containing ID and sequences to a dataframe
}
