% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_XML.R
\name{template_XML}
\alias{template_XML}
\title{Import template XML file}
\usage{
template_XML(file)
}
\arguments{
\item{file}{template XML file}
}
\description{
Take a template XML file and replace sequences and dates
}
